Uploading working CMSIS example

+++ TFLite_SAMD51/TFLite_SAMD51/main_files/TFLite_inference.h
@@ -11,4 +11,4 @@
 
 void setup_inference();
 
-float *run_inference();
+int8_t *run_inference();
